Rework Link `blank` for external links
The previous solution worked, but just by chance because Gatsby Link somehow captured the target=blank and ignored it.

This changed the default for external links only. We can still set it explicitly to clarify the behavior in components which is especially useful when the href does not show that it is external right in the code.